Update README.md
I wanted to include instruction that’s necessary for the user running the application to do which installs the dependent Node.js modules that are set forth in the package.json of each server on the application. This might be obvious to a developer with experience, but I always like to pretend that I’m writing instructions for someone who’s doing it for the first time, and this step felt necessary in hindsight to get added.